+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-##
+## 0.5.2
 - Add DNS server in AP mode ([#100](https://github.com/OpenRemise/Firmware/issues/100))
 - Add CSS to captive portal ([#103](https://github.com/OpenRemise/Firmware/pull/103))
+- Disable ZIMO MX decoder updates and uploads ([#105](https://github.com/OpenRemise/Firmware/issues/105))
 
 ## 0.5.1
 - Increase ADC sample rate to 80kHz